requirements.txt
flask==3.0.3
python-dotenv==1.0.1

aiogram==3.5.0

SQLAlchemy==2.0.32
